requirements.txt
streamlit
pandas
matplotlib
shap
fpdf2
numpy